refactor(OutputView): 출력을 담당하는 View 클래스 생성
