+++ src/main/java/lotto/Application.java
@@ -3,11 +3,13 @@ package lotto;
 import camp.nextstep.edu.missionutils.Console;
 import lotto.view.InputView;
 import lotto.controller.LottoController;
+import lotto.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
         LottoController lottoController = new LottoController(
-                new InputView()
+                new InputView(),
+                new OutputView()
         );
         lottoController.run();
         Console.close();
